% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sas_to_rds.R
\name{batch_sas_to_rds}
\alias{batch_sas_to_rds}
\title{Convert a batch of SAS files to RDS}
\usage{
batch_sas_to_rds(
  files = c("sample", "colonoscopy", "labs_a1c", "labs_bmp", "labs_cbc", "labs_crp",
    "labs_lft", "labs_lipid", ),
  ...
)
}
\arguments{
\item{files}{}

\item{...}{}
}
\value{

}
\description{
Convert a batch of SAS files to RDS
}
\seealso{
\code{\link{sas_to_rds}}
}
