% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysdata.R
\name{charlson_icd}
\alias{charlson_icd}
\title{ICD dictionary}
\usage{
charlson_icd(charl, icd = "icd9")
}
\arguments{
\item{charl}{}

\item{icd}{which icd code ("icd9" or "icd10")}
}
\value{
a function which maps a code to a 0/1 indicator if that code corresponds to the outcome charl
}
\description{
ICD dictionary
}
\keyword{internal}
