% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysdata.R
\docType{data}
\name{charlson_vars}
\alias{charlson_vars}
\title{Charlson variable names in new format}
\format{
An object of class \code{character} of length 16.
}
\usage{
charlson_vars
}
\value{

}
\description{
Charlson variable names in new format
}
\keyword{internal}
