% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_utils.R
\name{classification_metrics}
\alias{classification_metrics}
\title{Title}
\usage{
classification_metrics(pred, y, threshold = NULL)
}
\arguments{
\item{pred}{}

\item{y}{}

\item{threshold}{}
}
\description{
Title
}
