% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{create_charlson_data}
\alias{create_charlson_data}
\title{Title}
\usage{
create_charlson_data(
  dir = "./unzipped_data/",
  files = c("alldxs.sas7bdat"),
  which = charlson_vars,
  master = NULL,
  verbose = T
)
}
\arguments{
\item{dir}{}

\item{files}{}

\item{which}{}

\item{master}{}

\item{verbose}{}
}
\value{

}
\description{
Title
}
