% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{create_data}
\alias{create_data}
\title{Title}
\usage{
create_data(
  dir_in = "R_data/",
  file_out = "R_data/processed_records_5-1",
  y0 = -5,
  y1 = -1
)
}
\arguments{
\item{y0}{}

\item{y1}{}

\item{dir}{}
}
\value{

}
\description{
Title
}
