% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{create_event_data}
\alias{create_event_data}
\title{Title}
\usage{
create_event_data(
  dir = "./unzipped_data/",
  which = event_vars(),
  master = NULL
)
}
\arguments{
\item{dir}{}

\item{which}{}

\item{master}{}
}
\value{

}
\description{
Title
}
