% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{create_lab_data}
\alias{create_lab_data}
\title{Title}
\usage{
create_lab_data(dir = "./unzipped_data/", which = lab_types(), master = NULL)
}
\arguments{
\item{dir}{}

\item{which}{}

\item{master}{}
}
\value{

}
\description{
Title
}
