% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{create_meds_data}
\alias{create_meds_data}
\title{#' Title
#'
#' @param dir
#' @param which
#' @param master
#'
#' @return
#' @export
#' @import dplyr magrittr
#' @importFrom purrr reduce
#'
#' @examples
create_event_data = function(dir="./unzipped_data/", which=event_vars(), master=NULL, verbose=T){
dfs = list()
for(file in which){
if(verbose) cat(paste0("- ", file, " ...\n"))
src_df = load_sas(paste0(dir, file, ".sas7bdat"), file)
# restrict to prediction window
src_df \%<>\% left_join(master, by="ID")
cols = colnames(src_df); cols\link{2} = "date"; colnames(src_df) = cols
src_df \%<>\% filter((date>=start)&(date<=end))
# ensure ordered
src_df \%<>\% arrange(ID, date)
# lagged variables
src_df \%<>\% mutate(ID_lag = lag(ID), date_lag=lag(date))
new_subject = with(src_df, ID!=ID_lag)
src_df$ID_lag\link{new_subject} = NA; src_df$date_lag\link{new_subject} = NA
# difference and slope
src_df = src_df \%>\%
mutate(ddate=ifelse(date-date_lag==0, NA, date-date_lag)) \%>\%
mutate(sdate=1/ddate)
tmp = src_df \%>\% group_by(ID) \%>\% summarize(
n=n(),
maxdiff=safe_max(sdate)
)
colnames(tmp) = c("ID", paste(file, c("n", "maxdiff"), sep="_"))
dfs[\link{file}] = tmp
if(verbose) cat("  ...done.\n")
}
dff = dfs \%>\% purrr::reduce(full_join, by="ID")
return(dff)
}
Title}
\usage{
create_meds_data(
  dir = "./unzipped_data/",
  files = c("allmeds.sas7bdat"),
  which = med_vars,
  master = NULL,
  verbose = T
)
}
\arguments{
\item{dir}{}

\item{files}{}

\item{which}{}

\item{master}{}

\item{verbose}{}
}
\value{

}
\description{
#' Title
#'
#' @param dir
#' @param which
#' @param master
#'
#' @return
#' @export
#' @import dplyr magrittr
#' @importFrom purrr reduce
#'
#' @examples
create_event_data = function(dir="./unzipped_data/", which=event_vars(), master=NULL, verbose=T){
dfs = list()
for(file in which){
if(verbose) cat(paste0("- ", file, " ...\n"))
src_df = load_sas(paste0(dir, file, ".sas7bdat"), file)
# restrict to prediction window
src_df \%<>\% left_join(master, by="ID")
cols = colnames(src_df); cols\link{2} = "date"; colnames(src_df) = cols
src_df \%<>\% filter((date>=start)&(date<=end))
# ensure ordered
src_df \%<>\% arrange(ID, date)
# lagged variables
src_df \%<>\% mutate(ID_lag = lag(ID), date_lag=lag(date))
new_subject = with(src_df, ID!=ID_lag)
src_df$ID_lag\link{new_subject} = NA; src_df$date_lag\link{new_subject} = NA
# difference and slope
src_df = src_df \%>\%
mutate(ddate=ifelse(date-date_lag==0, NA, date-date_lag)) \%>\%
mutate(sdate=1/ddate)
tmp = src_df \%>\% group_by(ID) \%>\% summarize(
n=n(),
maxdiff=safe_max(sdate)
)
colnames(tmp) = c("ID", paste(file, c("n", "maxdiff"), sep="_"))
dfs[\link{file}] = tmp
if(verbose) cat("  ...done.\n")
}
dff = dfs \%>\% purrr::reduce(full_join, by="ID")
return(dff)
}
Title
}
