% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{HOSEA.fit}
\alias{HOSEA.fit}
\title{Title}
\usage{
HOSEA.fit(
  df,
  master,
  outcome = "ANY",
  imputation = "srs",
  to_drop = c("chol", "rbc", "hgb"),
  nrounds = 2000,
  print_every_n = 10,
  early_stopping_rounds = 100,
  ...
)
}
\arguments{
\item{df}{}

\item{master}{}

\item{outcome}{}

\item{imputation}{}

\item{to_drop}{}

\item{nrounds}{}

\item{print_every_n}{}

\item{early_stopping_rounds}{}

\item{...}{}
}
\description{
Title
}
