% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.R
\name{HOSEA.mice}
\alias{HOSEA.mice}
\title{Title}
\usage{
HOSEA.mice(df, vars_to_impute, n_rounds = 25, n_samples = 1, return_models = T)
}
\arguments{
\item{df}{}

\item{vars_to_impute}{}

\item{n_rounds}{}

\item{n_samples}{}

\item{return_models}{}
}
\value{
either the imputed df or a list of the df and the models
}
\description{
Title
}
