% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hosea.R
\docType{package}
\name{HOSEA}
\alias{HOSEA}
\title{HOSEA: Harnessing Opportunities to Screen for Esophageal Adenocarcinoma}
\description{
This package contains functions to process data, obtain predicted risks
and perform some analyses in relation to the HOSEA project.
}
\author{
Simon Fontaine \email{simfont@umich.edu}
}
