% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srs.R
\name{HOSEA.srs.fit}
\alias{HOSEA.srs.fit}
\title{Title}
\usage{
HOSEA.srs.fit(df, vars_to_impute, n_quantiles = 10000)
}
\arguments{
\item{df}{}

\item{vars_to_impute}{}

\item{n_quantiles}{}
}
\value{
the list of models
}
\description{
Title
}
