% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.R
\name{impute.HOSEA.mice}
\alias{impute.HOSEA.mice}
\title{Title}
\usage{
\method{impute}{HOSEA.mice}(obj, df, n_samples = 1, n_rounds = obj$n_rounds, cluster = NULL, ...)
}
\arguments{
\item{obj}{}

\item{df}{}

\item{n_samples}{}

\item{n_rounds}{}

\item{...}{}
}
\value{
a df or a list of dfs
}
\description{
Title
}
