% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srs.R
\name{impute.HOSEA.srs}
\alias{impute.HOSEA.srs}
\title{Title}
\usage{
\method{impute}{HOSEA.srs}(obj, df, n_samples = 1, ...)
}
\arguments{
\item{obj}{}

\item{df}{}

\item{n_samples}{}

\item{...}{}
}
\value{
a df or a list of dfs
}
\description{
Title
}
