% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysdata.R
\docType{data}
\name{lab_summaries}
\alias{lab_summaries}
\title{Lab measurement summaries names}
\format{
An object of class \code{character} of length 6.
}
\usage{
lab_summaries
}
\value{

}
\description{
Lab measurement summaries names
}
\keyword{internal}
