% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysdata.R
\docType{data}
\name{lab_vars}
\alias{lab_vars}
\title{Lab variables to keep}
\format{
An object of class \code{character} of length 29.
}
\usage{
lab_vars
}
\value{

}
\description{
Lab variables to keep
}
\keyword{internal}
