% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{load_models}
\alias{load_models}
\title{Title}
\usage{
load_models(
  files_meta = list(ANY = "xgb_srs_any.meta", EAC = "xgb_srs_eac.meta", GCA =
    "xgb_srs_egjac.meta"),
  files_models = list(ANY = "xgb_srs_any.model", EAC = "xgb_srs_eac.model", GCA =
    "xgb_srs_egjac.model")
)
}
\arguments{
\item{files_meta}{}

\item{files_models}{}
}
\description{
Title
}
