% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sas.R
\name{load_sas}
\alias{load_sas}
\title{Read a SAS file and perform some minor pre-processing}
\usage{
load_sas(filepath, filename, verbose = T, ...)
}
\arguments{
\item{filepath}{Path to the file to read}

\item{filename}{Name of the file}

\item{verbose}{How much to print}

\item{...}{Arguments to pass to read_sas}
}
\value{
the loaded data frame
}
\description{
When loading a SAS file, a RDS copy is made for future faster loading. If a RDS
file already exists with the same name, then this file is loaded instead.
}
