% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysdata.R
\docType{data}
\name{med_summaries}
\alias{med_summaries}
\title{Med summaries}
\format{
An object of class \code{character} of length 5.
}
\usage{
med_summaries
}
\value{

}
\description{
Med summaries
}
\keyword{internal}
