% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysdata.R
\name{other_vars}
\alias{other_vars}
\title{Other clinical variable names (after transformations)}
\usage{
other_vars()
}
\value{

}
\description{
Other clinical variable names (after transformations)
}
\keyword{internal}
