% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_utils.R
\name{patch_outcome}
\alias{patch_outcome}
\title{Title}
\usage{
patch_outcome(df, master, outcome = "ANY", drop = F)
}
\arguments{
\item{df}{}

\item{master}{}

\item{outcome}{}

\item{drop}{whether to drop other cases when one type is chosen}
}
\description{
Title
}
