% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_utils.R
\name{patch_staging}
\alias{patch_staging}
\title{Title}
\usage{
patch_staging(
  master,
  staging = paste0(system.file("extdata", package = "HOSEA"), "/staging.csv")
)
}
\arguments{
\item{master}{}

\item{staging}{}
}
\description{
Title
}
