% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.HOSEA}
\alias{predict.HOSEA}
\title{Obtain prediction from XGBoost model(s)}
\usage{
\method{predict}{HOSEA}(
  df,
  n_imputations = 10,
  xgb_meta = list(ANY = XGB_ANY, EAC = XGB_EAC, EGJAC = XGB_EGJAC),
  xgb_models = list(ANY = "", EAC = "", EGJAC = "")
)
}
\arguments{
\item{df}{data frame to predict}

\item{n_imputations}{number of imputation}

\item{xgb_fits}{a list of list(xgb_fit, quantiles, ...)}
}
\value{
predicted risk averaged over the n imputations. A data frame with columns (id, ANY, EAC, EGJAC).
}
\description{
Obtain prediction from XGBoost model(s)
}
