% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\title{Obtain prediction from XGBoost model}
\usage{
predict(xgb_fit, df, quantiles, n_imputations = 10)
}
\arguments{
\item{xgb_fit}{the xgboost object}

\item{df}{data frame to predict}

\item{quantiles}{quantiles for imputation}

\item{n_imputations}{number of imputation}
}
\value{
predicted risk averaged over the n imputations
}
\description{
Obtain prediction from XGBoost model
}
