% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{quality_control}
\alias{quality_control}
\title{Performs some simple checks on processed data}
\usage{
quality_control(df, model = XGB_ANY)
}
\arguments{
\item{df}{the dataframe to check}

\item{model}{which model to use for quantiles and missing proportions}
}
\value{
nothing, but will raise warnings when it finds potential issues.
}
\description{
Performs some simple checks on processed data
}
