% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_utils.R
\name{representative_sample}
\alias{representative_sample}
\title{Title}
\usage{
representative_sample(
  df,
  ratio_male_to_female = 1,
  ratio_cases_male_to_female = 8.33
)
}
\arguments{
\item{df}{}

\item{ratio_male_to_female}{}

\item{ratio_cases_male_to_female}{}
}
\description{
Title
}
