% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sas_to_rds.R
\name{sas_to_rds}
\alias{sas_to_rds}
\title{Import SAS data and convert to RDS}
\usage{
sas_to_rds(
  path_in = "unzipped_data/",
  path_out = "R_data/",
  filename = NULL,
  format_in = ".sas7dbat",
  format_out = ".rds"
)
}
\arguments{
\item{path_in}{}

\item{path_out}{}

\item{filename}{}

\item{format_in}{}

\item{format_out}{}
}
\description{
Import SAS data and convert to RDS
}
