% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{XGB_ANY}
\alias{XGB_ANY}
\title{XGBoost model trained on any cancer}
\format{
a list (xgb_fit, quantiles, missing_prop, test_ids)
}
\source{
HOSEA
}
\usage{
XGB_ANY
}
\description{
XGBoost model trained on any cancer
}
\keyword{datasets}
