% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{XGB_EAC}
\alias{XGB_EAC}
\title{XGBoost model trained on EAC cancers}
\format{
a list (xgb_fit, quantiles, missing_prop, test_ids)
}
\source{
HOSEA
}
\usage{
XGB_EAC
}
\description{
XGBoost model trained on EAC cancers
}
\keyword{datasets}
