% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{XGB_EGJAC}
\alias{XGB_EGJAC}
\title{XGBoost model trained on EGJAC cancers}
\format{
a list (xgb_fit, quantiles, missing_prop, test_ids)
}
\source{
HOSEA
}
\usage{
XGB_EGJAC
}
\description{
XGBoost model trained on EGJAC cancers
}
\keyword{datasets}
