% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{create_charlson_data}
\alias{create_charlson_data}
\title{Title}
\usage{
create_charlson_data(
  dir = "./unzipped_data/",
  prefix = "alldxs",
  which = charlson_names(),
  master = NULL
)
}
\arguments{
\item{dir}{}

\item{prefix}{}

\item{which}{}

\item{master}{}
}
\value{
a tibble with columns (ID, *which); with 0/1 entries depending whether
there was an ICD 9/10 code (see charlson_names()) during the prediction
interval (given by master)
}
\description{
Title
}
