% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{create_data}
\alias{create_data}
\title{Title}
\usage{
create_data(
  dir = "unzipped_data/",
  file_out = "R_data/processed_records/5-1.rds",
  y0 = -5,
  y1 = -1
)
}
\arguments{
\item{dir}{directory containing all files}

\item{y0}{starting year relative to index (default: -5)}

\item{y1}{ending year relative to index (default: -1)}
}
\value{

}
\description{
Title
}
