% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{create_lab_data}
\alias{create_lab_data}
\title{Title}
\usage{
create_lab_data(
  dir = "./unzipped_data/",
  files = c("alllabs.sas7bdat"),
  which = lab_vars,
  master = NULL,
  verbose = T
)
}
\arguments{
\item{dir}{}

\item{which}{}

\item{master}{}
}
\value{

}
\description{
Title
}
