% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{create_meds_data}
\alias{create_meds_data}
\title{Title}
\usage{
create_meds_data(dir = "./unzipped_data/", which = med_vars(), master = NULL)
}
\arguments{
\item{dir}{}

\item{which}{}

\item{master}{}
}
\value{

}
\description{
Title
}
