% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.R
\name{HOSEA.mice.fit}
\alias{HOSEA.mice.fit}
\title{Title}
\usage{
HOSEA.mice.fit(df, vars_to_impute, n_rounds = 10, cluster = NULL)
}
\arguments{
\item{df}{}

\item{vars_to_impute}{}

\item{n_rounds}{}

\item{cluster}{}
}
\value{
the list of models
}
\description{
Title
}
