% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srs.R
\name{HOSEA.srs}
\alias{HOSEA.srs}
\title{Title}
\usage{
HOSEA.srs(
  df,
  vars_to_impute,
  n_quantiles = 10000,
  n_samples = 1,
  return_models = T
)
}
\arguments{
\item{df}{}

\item{vars_to_impute}{}

\item{n_quantiles}{}

\item{n_samples}{}

\item{return_models}{}
}
\value{
either the imputed df or a list of the df and the models
}
\description{
Title
}
