% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srs.R
\name{impute.HOSEA.srs}
\alias{impute.HOSEA.srs}
\title{Impute using Simple Random Sampling from a fitted SRS imputer}
\usage{
\method{impute}{HOSEA.srs}(obj, df, n_samples = 1, ...)
}
\arguments{
\item{obj}{The SRS imputer}

\item{df}{The data frame to impute}

\item{n_samples}{the number of imputes to output}

\item{...}{For compatibility with \link{impute}}
}
\value{
a df or a list of dfs
}
\description{
Impute using Simple Random Sampling from a fitted SRS imputer
}
\seealso{
\link{HOSEA.srs} \link{HOSEA.srs.fit} \link{impute}
}
