% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{load_process_data}
\alias{load_process_data}
\title{Processing raw data to final format}
\usage{
load_process_data(
  dir = "unzipped_data/",
  files_sample = c("sample.sas7bdat"),
  files_charlson = c("alldxs.sas7bdat"),
  files_labs = c("alllabs.sas7bdat"),
  files_meds = c("allmeds.sas7bdat"),
  start = -4,
  end = 0,
  verbose = 2,
  icd = "any",
  icd10startdate = 17229,
  icd9enddate = 17229 - 3 * 31
)
}
\arguments{
\item{dir}{directory containing all files}

\item{files_sample}{list of files containing demographic data}

\item{files_charlson}{list of files containing icd codes}

\item{files_labs}{list of files containing lab results}

\item{files_meds}{list of files containing medication}

\item{start}{starting year relative to index (default: -4)}

\item{end}{ending year relative to index (default: -0)}

\item{verbose}{how much to print: 0=none, 1=some, 2=most, 3=all (default)}

\item{icd}{which icd codes to use (default: "any")}

\item{icd10startdate}{index of beginning of icd10 coding}

\item{icd9enddate}{index of end of icd9 coding}
}
\value{
a data frame (df) containing all required for fitting and/or prediction and a data frame (master) contaiing some metadata
}
\description{
Processing raw data to final format
}
