% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{load_process_data}
\alias{load_process_data}
\title{Title}
\usage{
load_process_data(
  dir = "unzipped_data/",
  start = -5,
  end = -1,
  icd10 = F,
  icd10startdate = 17229 - 3 * 365
)
}
\arguments{
\item{dir}{directory containing all files}

\item{start}{starting year relative to index (default: -5)}

\item{end}{ending year relative to index (default: -1)}

\item{file_out}{not used}
}
\value{

}
\description{
Title
}
