% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sas.R
\name{load_sas}
\alias{load_sas}
\title{Load sas files into R}
\usage{
load_sas(filepath, filename, verbose = T, ...)
}
\arguments{
\item{filepath}{}

\item{filename}{}

\item{...}{further arguments to pass to haven::read_sas}
}
\value{

}
\description{
Load sas files into R
}
