% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{patch_outcome}
\alias{patch_outcome}
\title{Title}
\usage{
patch_outcome(df, dir = "./unzipped_data/", which = "any")
}
\arguments{
\item{df}{}

\item{dir}{}

\item{which}{}
}
\value{

}
\description{
Title
}
