% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.HOSEA}
\alias{predict.HOSEA}
\title{Title}
\usage{
\method{predict}{HOSEA}(
  newdata,
  models = load_models(),
  imputer = load_imputer(),
  n_samples = 10
)
}
\arguments{
\item{newdata}{data frame to predict, which may contain missing data to be imputed}

\item{models}{a list of models to obtain predictions}

\item{imputer}{an imputation object from HOSEA, NULL to perform no imputation}

\item{n_samples}{}
}
\value{
predicted risk averaged over the n imputations. A data frame with columns (id, ANY, EAC, EGJAC).
}
\description{
Title
}
