% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.HOSEA}
\alias{predict.HOSEA}
\title{Obtain prediction from XGBoost model(s)}
\usage{
\method{predict}{HOSEA}(
  df,
  n_imputations = 10,
  xgb_meta = list(ANY = XGB_ANY, EAC = XGB_EAC, EGJAC = XGB_EGJAC),
  xgb_models = list(ANY = "xgb_any.model", EAC = "xgb_eac.model", EGJAC =
    "xgb_egjac.model"),
  use_json = T
)
}
\arguments{
\item{df}{data frame to predict}

\item{n_imputations}{number of imputation (default:10)}

\item{xgb_meta}{list of objects with quantiles (default: all three models)}

\item{xgb_models}{list of file names to load models (default: all three models)}

\item{use_json}{whether to use the json model or not (default:T)}
}
\value{
predicted risk averaged over the n imputations. A data frame with columns (id, ANY, EAC, EGJAC).
}
\description{
Obtain prediction from XGBoost model(s)
}
