% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\title{Obtain prediction from XGBoost model(s)}
\usage{
predict(
  df,
  n_imputations = 10,
  xgb_fits = list(ANY = XGB_1M_ANY, EAC = XGB_1M_EAC, EGJAC = XGB_1M_EGJAC)
)
}
\arguments{
\item{df}{data frame to predict}

\item{n_imputations}{number of imputation}
}
\value{
predicted risk averaged over the n imputations. A data frame with columns (id, ANY, EAC, EGJAC).
}
\description{
Obtain prediction from XGBoost model(s)
}
